feat(posts): apply vertical slice to posts feature
