+++ src/posts/domain/interfaces/posts-repository.ts
@@ -1,5 +1,5 @@
-import type { CreatePostInputModel } from "@/src/core/posts/domain/models/create-post-input-model";
-import type { Post } from "@/src/core/posts/domain/models/post";
+import type { CreatePostInputModel } from "@/src/posts/domain/models/create-post-input-model";
+import type { Post } from "@/src/posts/domain/models/post";
 import type { Page } from "@/src/core/app/domain/models/page";
 
 export interface IPostsRepository {
