+++ src/posts/domain/models/create-post-input-model.ts
