+++ src/posts/domain/models/post.ts
