+++ src/posts/presentation/pages/posts/posts.tsx
@@ -1,13 +1,13 @@
-import css from "../../../users/views/users-page/users-page.css";
+import css from "@/src/common/presentation/styles/wrapper.css";
 import { SimpleCard } from "@/src/ui/components/simple-card/simple-card";
 import { locator } from "@/src/core/app/ioc/__generated__";
 import type { IocProvider } from "@/src/core/app/ioc/interfaces";
-import type { GetPostsUseCase } from "@/src/core/posts/domain/use-cases/get-posts-use-case";
+import type { GetPostsUseCase } from "@/src/posts/domain/use-cases/get-posts-use-case";
 import { TYPES } from "@/src/core/app/ioc/__generated__/types";
-import PageTitle from "@/src/ui/features/posts/views/posts-page/components/page-title";
-import type { Post } from "@/src/core/posts/domain/models/post";
+import type { Post } from "@/src/posts/domain/models/post";
 import { notFound } from "next/navigation";
 import { getTranslations } from "next-intl/server";
+import PostsPageTitle from "@/src/posts/presentation/components/posts-page-title";
 
 export default async function PostsPage() {
   let posts: Array<Post> = [];
@@ -21,7 +21,7 @@ export default async function PostsPage() {
 
   return (
     <div className={css.wrapper}>
-      <PageTitle title={t("title")} subtitle={t("subtitle")} />
+      <PostsPageTitle title={t("title")} subtitle={t("subtitle")} />
       {posts.map((post, idx) => (
         <SimpleCard key={`${post.id}_${idx}`} title={post.title} subtitle={post.body} />
       ))}
